% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_mapping_rates}
\alias{calculate_mapping_rates}
\title{Calculate Mapping Rates}
\usage{
calculate_mapping_rates(mapped_input_list, list_input_files, myinputfiles)
}
\arguments{
\item{mapped_input_list}{list of mapped input files}

\item{list_input_files}{list of input files}

\item{myinputfiles}{list of input files}
}
\value{
vector of mapping rates for each input file
}
\description{
Calculate Mapping Rates
}
\author{
Andrew Patt
}
