% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFunctions.R
\name{getChemPropsFromName}
\alias{getChemPropsFromName}
\title{getChemProps}
\usage{
getChemPropsFromName(db, nameList)
}
\arguments{
\item{db}{db connection}

\item{nameList}{input vector of metabolite names}
}
\description{
getChemProps
}
\author{
Patt
}
