% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeFunctions_ACP.R
\name{harmonizefiles}
\alias{harmonizefiles}
\title{Harmonizing metabolites across multiple files}
\usage{
harmonizefiles(fileList, filterOnlyOne)
}
\arguments{
\item{fileList}{list of dataframes of metabolites with RefMet query results}

\item{filterOnlyOne}{boolean for whether to include or exclude metabolites mapped in only one file}
}
\value{
dataframe of metabolites harmonized across all files
}
\description{
Harmonizing metabolites across multiple files
}
