% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeFunctions_ACP.R
\name{mapMetabolitesToRefMet}
\alias{mapMetabolitesToRefMet}
\title{Mapping metabolites from data to refmet for standardized names}
\usage{
mapMetabolitesToRefMet(input_df, filename, HMDB_col, metab_col, CID_col)
}
\arguments{
\item{input_df}{a dataframe with metabolite names, ids (HMDB, CID), information}

\item{filename}{name of file with data in input_df}

\item{HMDB_col}{name of column containing HMDB IDs}

\item{metab_col}{name of column containing metabolite names}

\item{CID_col}{name of column containing PubChem IDs}
}
\value{
list of dataframe of metabolite and information from RefMet and list of metabolites unharmonizable
}
\description{
Mapping metabolites from data to refmet for standardized names
}
\examples{
\dontrun{
metFiles <- mapMetabolitesToRefMet(input_df = metabolite_information,
filename = "Original File", HMDB_col = "HMDB_ids",
metab_col = "metab_name", CID_col = "pubchem")
}
}
