% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{massCheck}
\alias{massCheck}
\title{Mass Check}
\usage{
massCheck(synonym_DF)
}
\arguments{
\item{synonym_DF}{dataframe of synonyms}
}
\value{
dataframe with three possible values per rampId: species, class or invalid
}
\description{
Mass Check
}
\author{
Iris Pang, Andrew Patt
}
