% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserFunctions.R
\name{merge_files}
\alias{merge_files}
\title{Merge Files}
\usage{
merge_files(mapped_input_list, myinputfiles)
}
\arguments{
\item{mapped_input_list}{Mapped input list}

\item{myinputfiles}{Input files}
}
\value{
Merged files
}
\description{
Merge Files
}
