% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserFunctions.R
\name{parse_metabolon_names}
\alias{parse_metabolon_names}
\title{Parse Names}
\usage{
parse_metabolon_names(ids, RaMP_prefixes)
}
\arguments{
\item{ids}{A vector of metabolite ids}

\item{RaMP_prefixes}{A vector of RaMP prefixes}
}
\value{
A vector of metabolite names
}
\description{
Parse Names
}
\author{
Patt
}
