% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFunctions.R
\name{queryRampSynonyms}
\alias{queryRampSynonyms}
\title{Query Ramp Synonyms}
\usage{
queryRampSynonyms(ids, db = RaMP::RaMP(), use_metabolon_parsers = TRUE)
}
\arguments{
\item{ids}{a string of metabolites separated by newline}

\item{db}{connection to RaMP sqlite database}

\item{use_metabolon_parsers}{Strip special characters typical of metabolon inputs}
}
\value{
a dataframe of all RaMP synonyms found for inputs,
with class/species flags
}
\description{
Query Ramp Synonyms
}
\author{
Andrew Patt
}
