% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeFunctions_ACP.R
\name{queryRefMet}
\alias{queryRefMet}
\title{Query RefMet for standardized names}
\usage{
queryRefMet(mets)
}
\arguments{
\item{mets}{a string of metabolites separated by newline}
}
\value{
a dataframe of RefMet query results: input/standardized name, classes, molecular formula, mass
}
\description{
Query RefMet for standardized names
}
\examples{
\dontrun{
queryResults <- queryRefMet(mets = "2'-Deoxyuridine")
}
}
