% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserFunctions.R
\name{readInputCSVs}
\alias{readInputCSVs}
\title{Read Input CSVs}
\usage{
readInputCSVs(inputcsv)
}
\arguments{
\item{inputcsv}{character of pathway to csv containing names of files with metabolite names}
}
\value{
a list of input files
}
\description{
Read Input CSVs
}
