% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFunctions.R
\name{refmet_metadata}
\alias{refmet_metadata}
\title{refmet_metadata}
\usage{
refmet_metadata(metabolite)
}
\arguments{
\item{metabolite}{RefMet metabolite name}
}
\value{
metadata
}
\description{
refmet_metadata
}
\author{
Eoin Fahy
}
