% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeFunctions_ACP.R
\name{replaceEmptys}
\alias{replaceEmptys}
\title{Replace empty cells with NA}
\usage{
replaceEmptys(input_df)
}
\arguments{
\item{input_df}{a dataframe with empty cells}
}
\value{
a modified dataframe where all empty cells are replaced with NA
}
\description{
Replace empty cells with NA
}
\examples{
\dontrun{
updatedf <- replaceEmptys(input_df = data.frame(A = c("", "B"), B = c("C", "E")))
}
}
