% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizedFileforShare_EM2.R
\name{runHarmonization}
\alias{runHarmonization}
\title{Run total harmonization of metabolites}
\usage{
runHarmonization(inputcsv, outputFileName, writecsv)
}
\arguments{
\item{inputcsv}{character of pathway to csv containing names of files with metabolite names}

\item{outputFileName}{character of output file}

\item{writecsv}{boolean for whether or not to write a csv file of the dataframe with harmonized files}
}
\value{
a dataframe containing harmonized files across all input files listed in inputcsv
}
\description{
Run total harmonization of metabolites
}
\examples{
\dontrun{
finalHarmonized <- runHarmonization(inputcsv = "HarmInputFiles.csv",
outputFileName = "exampleharmonizedfile")
}
}
