% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFunctions.R
\name{send_id_vector_to_RefMet}
\alias{send_id_vector_to_RefMet}
\title{send_id_vector_to_RefMet}
\usage{
send_id_vector_to_RefMet(id_vector)
}
\arguments{
\item{id_vector}{ids to send to refMet (vector)}
}
\value{
Dataframe of RefMet results
}
\description{
send_id_vector_to_RefMet
}
\author{
Patt
}
