% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFunctions.R
\name{queryRefMet}
\alias{queryRefMet}
\title{Query RefMet for standardized names. Code from Eoin Fahy, received on 5/8/2023}
\usage{
queryRefMet(
  input_df,
  filename,
  HMDB_col,
  CID_col,
  KEGG_col = NA,
  LM_col = NA,
  CHEBI_col = NA,
  metab_col,
  synonym_search = FALSE
)
}
\arguments{
\item{input_df}{dataframe of metabolites}

\item{filename}{name of file}

\item{HMDB_col}{HMDB column name}

\item{CID_col}{CID column name}

\item{KEGG_col}{KEGG column name}

\item{LM_col}{LM column name}

\item{CHEBI_col}{CHEBI column name}

\item{metab_col}{metabolite column name}

\item{synonym_search}{boolean for whether to search synonyms
#' @return a dataframe of RefMet query results: input/standardized name, classes, molecular formula, mass}
}
\description{
Query RefMet for standardized names. Code from Eoin Fahy, received on 5/8/2023
}
\examples{
\dontrun{
queryResults <- queryRefMet(mets = "2'-Deoxyuridine")
}

}
